move PicaMassImport-Plugin into Kitodo
