+++ Kitodo/src/main/java/org/kitodo/production/plugin/importer/massimport/PicaMassImport.java
@@ -1,4 +1,4 @@
-package PicaMassImport.de.intranda.goobi.plugins;
+package org.kitodo.production.plugin.importer.massimport;
 
 /**
  * Copyright by intranda GmbH 2013. All rights reserved.
@@ -88,11 +88,11 @@ import ugh.exceptions.ReadException;
 import ugh.exceptions.TypeNotAllowedForParentException;
 import ugh.exceptions.WriteException;
 import ugh.fileformats.mets.MetsMods;
-import PicaMassImport.de.intranda.goobi.plugins.sru.SRUHelper;
+import org.kitodo.production.plugin.importer.massimport.sru.SRUHelper;
 import de.sub.goobi.beans.Prozesseigenschaft;
 import de.sub.goobi.beans.Vorlageeigenschaft;
 import de.sub.goobi.beans.Werkstueckeigenschaft;
-import PicaMassImport.de.sub.goobi.helper.UghUtils;
+import org.kitodo.production.plugin.importer.massimport.UghUtils;
 import de.sub.goobi.helper.exceptions.ImportPluginException;
 
 @PluginImplementation
@@ -422,7 +422,7 @@ public class PicaMassImport implements IImportPlugin, IPlugin {
 	 * @param field
 	 *            field value to return
 	 * @return field value, or ""
-	 * @see org.goobi.production.plugin.CataloguePlugin.PicaPlugin.PicaPlugin#getPpnFromParent(Element,
+	 * @see org.kitodo.production.plugin.opac.pica.PicaPlugin#getPpnFromParent(Element,
 	 *      String, String)
 	 */
 	@SuppressWarnings("unchecked")
@@ -446,7 +446,7 @@ public class PicaMassImport implements IImportPlugin, IPlugin {
 	 * @param attributeValue
 	 *            attribute to locate
 	 * @return value, or "" if not found
-	 * @see org.goobi.production.plugin.CataloguePlugin.PicaPlugin.PicaPlugin#getSubelementValue(Element,
+	 * @see org.kitodo.production.plugin.opac.pica.PicaPlugin#getSubelementValue(Element,
 	 *      String)
 	 */
 	@SuppressWarnings("unchecked")
@@ -803,7 +803,7 @@ public class PicaMassImport implements IImportPlugin, IPlugin {
 	}
 
 	/**
-	 * @param the opac catalogue
+	 * @param opacCatalogue the opac catalogue
 	 */
 	@Override
 	public void setOpacCatalogue(String opacCatalogue) {
@@ -818,7 +818,7 @@ public class PicaMassImport implements IImportPlugin, IPlugin {
 	}
 
 	/**
-	* @param the goobi config directory
+	* @param configDir the goobi config directory
 	*/
 	@Override
 	public void setGoobiConfigDirectory(String configDir) {
