+++ Kitodo/src/main/java/org/kitodo/production/plugin/importer/massimport/googlecode/fascinator/redbox/sru/SRUClient.java
@@ -16,7 +16,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package PicaMassImport.com.googlecode.fascinator.redbox.sru;
+package org.kitodo.production.plugin.importer.massimport.googlecode.fascinator.redbox.sru;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
